Sign for datacube assets (#28)
* Sign for datacube assets

This adds support for signing assets that point to Zarr datasets in
Azure Blob Storage. We look for assets with `xarray:storage_options`
and, if present, add a `credential` field with the read-only SAS token.

These can then be passed along to fsspec / xarray, as recommended in
https://planetarycomputer.microsoft.com/docs/quickstarts/reading-zarr-data/.

I also upped the test coverage for Zarr and tabular datasets.